requirements.txt
langfuse==2.58.2
pandas==2.2.3
pathlib==1.0.1
pip==22.3
pydantic==2.10.5
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.32.3
uvicorn==0.34.0
fastapi==0.115.7